+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Readiness probe added ([#241])
+
 ### Changed
 
 - BREAKING: The deep storage on s3 and the s3 config for ingestion have been changed to use the operator-rs commons::s3 structs ([#228])
@@ -12,6 +16,7 @@ All notable changes to this project will be documented in this file.
 
 [#228]: https://github.com/stackabletech/druid-operator/pull/228
 [#238]: https://github.com/stackabletech/druid-operator/pull/238
+[#241]: https://github.com/stackabletech/druid-operator/pull/241
 
 ## [0.5.0] - 2022-03-15